requirements.txt
#django-phonenumber-field[phonenumberslite]
#pillow
#Django==2.1.5
#anyjson==0.3.3
#django-crispy-forms==1.9.2
#validate_email

asgiref==3.3.1
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
Cities==1.1
cryptography==3.3.1
defusedxml==0.6.0
django-allauth==0.44.0
django-braces==1.14.0
django-location-field==2.1.0
django-oauth-toolkit==0.10.0
django-phonenumber-field==4.0.0
docker==4.4.1
idna==2.10
numpy==1.19.4
oauthlib==1.0.3
pandas==1.2.0
phonenumberslite==8.12.15
pycparser==2.20
PyJWT==2.0.0
PyMySQL==0.10.1
PyPrind==2.11.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
pywin32==227
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
websocket-client==0.57.0